correction: Splitting a block with one movie into two; Moved the logic of the user's fields to the UserContext; Now MoviesBlock only accepts movies; Pagination Black has been changed - the page counting function has been removed, now the first and last pages are displayed in numbers and only when they are not visible; A constant of 500 pages has been introduced; Icons are now in svg; In the files, simpleMovie/s is replaced by movie/s.
