+++ src/constants/movie_constants.ts
@@ -1,1 +1,2 @@
 export const MAX_MOVIES_ON_PAGE = 20;
+export const MAX_TOTAL_PAGES = 500;
