+++ src/index.tsx
@@ -3,12 +3,13 @@ import "./index.css";
 import { fillLocalStorage } from "./fillLocalStorage";
 import { I18nextProvider } from "react-i18next";
 import { RouterProvider } from "react-router-dom";
-import { AuthProvider } from "./global-components/AuthContext";
-import { LanguageProvider } from "./global-components/switch-language/LanguageContext";
+import { AuthProvider } from "./components/AuthContext";
+import { LanguageProvider } from "./components/switch-language/LanguageContext";
 import i18n from "./i18n";
 import router from "./routes";
-import NavBar from "./global-components/NavBar";
-import { ViewProvider } from "./components/movies-block/switch-view/ViewContext";
+import NavBar from "./components/NavBar";
+import { ViewProvider } from "./pages/home/movies-block/switch-view/ViewContext";
+import { UserProvider } from "./components/UserContext";
 
 fillLocalStorage();
 
@@ -20,8 +21,10 @@ root.render(
   <I18nextProvider i18n={i18n}>
     <AuthProvider>
       <LanguageProvider>
-        <NavBar />
-        <RouterProvider router={router} />
+        <UserProvider>
+          <NavBar />
+          <RouterProvider router={router} />
+        </UserProvider>
       </LanguageProvider>
     </AuthProvider>
   </I18nextProvider>
