+++ src/pages/home/GenresBlock.tsx
@@ -1,15 +1,16 @@
 import { FC, useState, useEffect } from "react";
-import { useAuth } from "../../global-components/AuthContext";
+import { useAuth } from "../../components/AuthContext";
 import { Genre } from "../../types/movie-api-types/Genre";
 import { useTranslation } from "react-i18next";
+import { useUser } from "../../components/UserContext";
 
 interface GenresBlockProps {
   genres: Genre[];
 }
 
 const GenresBlock: FC<GenresBlockProps> = ({ genres }) => {
-  const { user, addFavoriteGenre, removeFavoriteGenre, getFavoriteGenresId } =
-    useAuth();
+  const { user } = useAuth();
+  const { getFavoriteGenresId, removeFavoriteGenre, addFavoriteGenre } = useUser();
   const [favoriteGenresId, setFavoriteGenresId] = useState<number[]>(
     getFavoriteGenresId()
   );
@@ -33,6 +34,7 @@ const GenresBlock: FC<GenresBlockProps> = ({ genres }) => {
       <div className="mt-5 grid grid-cols-10 grid-rows-2 gap-2">
         {genres.map((genre) => (
           <button
+          key={genre.id}
             className={`hover:bg-blue-950 hover:text-white rounded-full px-5 py-2 overflow-hidden truncate ${
               favoriteGenresId.includes(genre.id)
                 ? "bg-white text-blue-950"
