+++ src/pages/home/movies-block/switch-view/SwitchViewButton.tsx
