+++ src/pages/home/movies-block/switch-view/ViewContext.tsx
