+++ src/routes.tsx
@@ -1,5 +1,5 @@
 import { createBrowserRouter } from "react-router-dom";
-import ProtectedRoute from "./global-components/ProtectedRoute";
+import ProtectedRoute from "./components/ProtectedRoute";
 import Home from "./pages/home/Home";
 import Login from "./pages/login/Login";
